Add DataFoodConsortium::Connector::SKOSHelper
It contains method needed by both SKOSInstance and SKOSConcept
